keeping home_loc / work_loc
Maintaining the home_loc/work_loc allows us to visualize the detected home/work across all days. Example: If in weeks [0, i] we detected loc 45 as Home, but in Week i+1 we do not visit home, we can still predict 45 as Home (thanks to the window aggregations). This is visible in home_loc, but not in locations_labels.